5217: Use working defaults for keystore path
To make it as easy as possible to get up and running with the eIDAS
reference nodes we'd like to make sure that the code has defaults which
will "just work", irrespective of the platform (*nix / windows) that
you're running on. To this effect we're adding a keystore to
`keystore/eidasKeystore` and updating the config to use relative paths
to this keystore instead of absolute paths to a keystore on the (windows
only) C: drive.

Ultimately we'll probably want to make these settings environment
variables, not just relative paths. But first we want to get it all
running.

NOTE: It still doesn't run yet...

Authors: @richardtowers @chrisholmes