5217: More environment variable aware config
Fixed a couple of new places that need to be aware of environment
variables in config. This is enough (I think) to make eidas.xml aware of
environment variables.

NOTE: There's still a test failing, because we're copying properties
instead of mutating them. TODO: fix this.

Authors: @niyioyelade @richardtowers @chrisholmes