5217: Switch to mutating properties files when substituting env vars
We were sneaking in at the setConfigs level. The existing code expects
setConfigs to set *the provided instance* of properties on the class,
not a copy of it.

It definitely feels super-grubby taking an already-loaded properties
file and mutating values like this. This might mean we need to
reconsider our approach here entirely.

One suggested approach is to have cloud foundry run a script that stamps
out some config file templates with environment variables before the
application starts. This would have the significant advantage that we
wouldn't have to make any changes to the source.

Authors: @richardtowers