5218: Remove all exclusions from pom files
Previously many transitive dependencies were being excluded in the maven
config. This was presumably done for compatability with a range of
application servers, some of which already bundle some of these files
(for example there was a comment saying "upsets jboss 6"). Unfortunately
excluding dependencies means we would need an extra step to copy these
jar files somewhere for tomcat, which is an extra bit of complexity we
could do without.

Since compatibility with other application servers is not a priority for
the UK at this point we're removing the exclusions and letting maven do
its thing.

There was one issue in the stub SP with slf4j that really did require an
exclusion (the tests were failing without it), so we've left this one
in.

Authors: @richardtowers @niyioyelade