5218: Copy EIDAS-Config into the .war files
When deploying to a PaaS environment it's much more convenient if all
the files needed for the application to run live in one place. In the
case of the config it's best if everything is inside the .war file.
We've already amended the application so that it can be configured using
environment variables, so the same war file can still be deployed to
different environments.

Authors: @richardtowers @niyioyelade