5218: Add a script to deploy reference implementations to PaaS
In the UK we have a Platform as a Service on which we're planning to
deploy the eIDAS reference implementations for our Alpha. This adds a
shell script to invoke the right cloud foundry commands to deploy the
nodes.

This is still a work in progress, not all of the applications start on
our PaaS yet.

Authors: @richardtowers @niyioyelade